dashv2/requirements.txt
boto==2.38.0
bottle==0.12.8
isodate==0.5.1
jira==0.47
lockfile==0.10.2
pyforce==1.6
pymongo==2.8
python-daemon==2.0.5
python-dateutil==2.4.2
elasticsearch==1.5.0
pytz==2015.4
pyyaml==3.11
requests==2.7.0
wsgiproxy==0.2.2
